stat tweaks sleeping eye and moved files
changed the .maxDamage of the tier 2 and 3 whips and knives, moved each whip tier into there own folder, added the sleeping eye of ender and removed the vanilla recipe for the eye of ender. Note: there is no recipe for the sleeping eye and no tier 2 and 3 textures in this commit.