new folders and fixes (#6)
* whip redo part 1

Changed all the whip textures added a new whip texture for tier 1-3 base whips and added a tier 1 red whip texture. Added the code for the Runic whip + knife and changed all the stats of the whips for more game balance

* Added the base whip T1-3 Textures

* Creative Menu fix + New Folders

Fixed the kubejs creative menu not being named 'Magical Addons' and missing the logo. Added the folder defaultconfigs and saves as in Structure Upgrader World there is the datapack for custom structures (based on Structory)